+++ chemics/chemical_equation.py
@@ -1,7 +1,6 @@
 import chemics as cm
-import numpy as np
+import pandas as pd
 import re
-import textwrap
 from collections import Counter
 
 
@@ -22,48 +21,26 @@ class ChemicalEquation:
         Chemical equation such as A + B -> C + D
     names : dict, optional
         Names of unique species and their corresponding chemical formula.
+    rct_properties : dataframe
+        Number of moles, chemical species, molecular weight, mass, mole
+        fraction, and mass fraction for each reactant.
+    rct_elements : dict
+        Total number of atomic elements on reactant side of the equation.
+    rct_moles : float
+        Total number of moles on reactant side of the equation.
+    rct_mass : float
+        Total mass of the reactants.
+    prod_properties : dataframe
+        Number of moles, chemical species, molecular weight, mass, mole
+        fraction, and mass fraction for each product.
+    prod_elements : dict
+        Total number of atomic elements on product side of the equation.
+    prod_moles : float
+        Total number of moles on product side of the equation.
+    prod_mass : float
+        Total mass of the products.
     balance : bool
         Balance between atomic elements in reactants and products.
-    reactants : list
-        Reactants in the chemical equation.
-    rct_species : list
-        Chemical species from reactants list. Uses the `names` dictionary if needed.
-    rct_moles : array
-        Number of moles for each chemical species in reactants.
-    rct_mw : array
-        Molecular weight for each chemical species in reactants.
-    rct_mass : array
-        Molecular weight for each chemical species in reactants.
-    rct_elements : dict
-        Number of atomic elements in reactants.
-    rct_sum_moles : float
-        Total number of moles for reactants.
-    rct_sum_mass : float
-        Total mass of reactants.
-    rct_mol_fracs : array
-        Mole fractions of each chemical species in reactants.
-    rct_mass_fracs : array
-        Mass fractions of each chemical species in reactants.
-    products : list
-        Products in the chemical equation.
-    prod_species : list
-        Chemical species from products list. Uses the `names` dictionary if needed.
-    prod_moles : array
-        Number of moles for each chemical species in products.
-    prod_mw : array
-        Molecular weight for each chemical species in products.
-    prod_mass : array
-        Molecular weight for each chemical species in products.
-    prod_elements : dict
-        Number of atomic elements in products.
-    prod_sum_moles : float
-        Total number of moles for products.
-    prod_sum_mass : float
-        Total mass of products.
-    prod_mol_fracs : array
-        Mole fractions of each chemical species in products.
-    prod_mass_fracs : array
-        Mass fractions of each chemical species in products.
 
     Examples
     --------
@@ -71,15 +48,30 @@ class ChemicalEquation:
     >>> ce.balance
     True
 
-    >>> ce.rct_mol_fracs
-    [0.5 0.5]
+    >>> ce.rct_properties
+                   HCl        Na
+    moles            2         2
+    species        HCl        Na
+    molwt       36.458     22.99
+    mass        72.916     45.98
+    molfrac        0.5       0.5
+    massfrac  0.613275  0.386725
 
-    >>> print(ce)
-    # this prints all attributes to the screen
+    >>> ce.rct_properties.loc['massfrac']
+    HCl    0.613275
+    Na     0.386725
+
+    >>> ce.rct_elements
+    {'H': 2.0, 'Cl': 2.0, 'Na': 2.0}
+
+    >>> ce.rct_moles
+    4.0
+
+    >>> ce.rct_mass
+    118.896
     """
 
     def __init__(self, eq, names=None):
-
         self.eq = eq
         self.names = names
         self._parse_equation()
@@ -87,86 +79,59 @@ class ChemicalEquation:
         self._assign_prod_attrs()
         self._check_balance()
 
-    def __str__(self):
-
-        desc = f"""
-        {'eq':18} {self.eq}
-        {'names':18} {self.names}
-        {'balance':18} {self.balance}
-
-        {'reactants':18} {self.reactants}
-        {'rct_species':18} {self.rct_species}
-        {'rct_moles':18} {self.rct_moles}
-        {'rct_mw':18} {self.rct_mw}
-        {'rct_mass':18} {self.rct_mass}
-        {'rct_elements':18} {self.rct_elements}
-        {'rct_sum_moles':18} {self.rct_sum_moles}
-        {'rct_sum_mass':18} {self.rct_sum_mass}
-        {'rct_mol_fracs':18} {self.rct_mol_fracs}
-        {'rct_mass_fracs':18} {self.rct_mass_fracs}
-
-        {'products':18} {self.products}
-        {'prod_species':18} {self.prod_species}
-        {'prod_moles':18} {self.prod_moles}
-        {'prod_mw':18} {self.prod_mw}
-        {'prod_mass':18} {self.prod_mass}
-        {'prod_elements':18} {self.prod_elements}
-        {'prod_sum_moles':18} {self.prod_sum_moles}
-        {'prod_sum_mass':18} {self.prod_sum_mass}
-        {'prod_mol_fracs':18} {self.prod_mol_fracs}
-        {'prod_mass_fracs':18} {self.prod_mass_fracs}"""
-
-        return textwrap.dedent(desc)
-
     def _parse_equation(self):
         """
         Split chemical equation into reactants and products.
         """
-        rcts = self.eq.split('->')[0].split('+')
-        self.reactants = [r.strip() for r in rcts]
+        rct_split = self.eq.split(' -> ')[0].split(' + ')
+        self._rct_items = rct_split
 
-        prods = self.eq.split('->')[1].split('+')
-        self.products = [p.strip() for p in prods]
+        prod_split = self.eq.split(' -> ')[1].split(' + ')
+        self._prod_items = prod_split
 
-    def _equation_properties(self, comps):
+    def _eq_properties(self, eq_items):
         """
-        Determine properties for each chemical component in reactants or
-        products.
+        Determine properties for each item in reactants or products.
         """
-        nc = len(comps)
-        element_counter = Counter()
+        eq_names = []
+        eq_moles = []
+        eq_species = []
+        eq_molwts = []
+        eq_masses = []
+        eq_elements = Counter()
 
-        species = []
-        moles = np.zeros(nc)
-        molwt = np.zeros(nc)
-        mass = np.zeros(nc)
+        for item in eq_items:
 
-        for i, c in enumerate(comps):
-
-            # --- moles and chemical species ---
-
-            if c[0].isdigit():
-                c = c.split()
-                mol = float(c[0])
-                sp = c[1]
+            # number of moles and name for each item
+            if item[0].isdigit():
+                item = item.split()
+                mol = float(item[0])
+                name = item[1]
+                eq_names.append(name)
+                eq_moles.append(mol)
             else:
                 mol = 1.0
-                sp = c
+                name = item
+                eq_names.append(name)
+                eq_moles.append(mol)
 
-            if (self.names is not None) and (sp in self.names.keys()):
-                sp = self.names[sp]
+            # species for each item is chemical formula specified in names
+            # dictonary or as chemical formula (name) in the chemical equation
+            # note - species is used for the molecular weight and elements
+            if (self.names is not None) and (name in self.names.keys()):
+                sp = self.names[name]
+                eq_species.append(sp)
+            else:
+                sp = name
+                eq_species.append(sp)
 
-            species.append(sp)
-            moles[i] = mol
-
-            # --- mass and molecular weight ---
-
+            # masses and molecular weights for each item
             mw = cm.mw(sp)
-            molwt[i] = mw
-            mass[i] = mol * mw
+            mass = mol * mw
+            eq_molwts.append(cm.mw(sp))
+            eq_masses.append(mass)
 
-            # --- count elements from each chemical species ---
-
+            # count elements from each chemical species
             rex = re.findall('([A-Z][a-z]?)([0-9]*)', sp)
 
             for r in rex:
@@ -179,55 +144,49 @@ class ChemicalEquation:
 
                 # dict where key is element and value is number of atoms such as {'C': 6.0}
                 d = {element: atoms * mol}
-                element_counter.update(d)
+                eq_elements.update(d)
 
-        sum_moles = sum(moles)
-        sum_mass = sum(mass)
-        mol_fracs = moles / sum_moles
-        mass_fracs = mass / sum_mass
+        # sum of moles and masses
+        eq_sum_moles = sum(eq_moles)
+        eq_sum_masses = sum(eq_masses)
 
-        properties = dict(
-            species=species,
-            moles=moles,
-            molwt=molwt,
-            mass=mass,
-            element_counter=element_counter,
-            sum_moles=sum_moles,
-            sum_mass=sum_mass,
-            mol_fracs=mol_fracs,
-            mass_fracs=mass_fracs)
+        # mole fractions and mass fractions
+        eq_mol_fracs = [m / eq_sum_moles for m in eq_moles]
+        eq_mass_fracs = [m / eq_sum_masses for m in eq_masses]
 
-        return properties
+        # dataframe for properties
+        cols = eq_names
+        idx = ['moles', 'species', 'molwt', 'mass', 'molfrac', 'massfrac']
+
+        df = pd.DataFrame(columns=cols, index=idx)
+        df.loc['moles'] = eq_moles
+        df.loc['species'] = eq_species
+        df.loc['molwt'] = eq_molwts
+        df.loc['mass'] = eq_masses
+        df.loc['molfrac'] = eq_mol_fracs
+        df.loc['massfrac'] = eq_mass_fracs
+
+        return df, dict(eq_elements), eq_sum_moles, eq_sum_masses
 
     def _assign_rct_attrs(self):
         """
         Assign results for reactants to attributes.
         """
-        rct_properties = self._equation_properties(self.reactants)
-        self.rct_species = rct_properties['species']
-        self.rct_moles = rct_properties['moles']
-        self.rct_mw = rct_properties['molwt']
-        self.rct_mass = rct_properties['mass']
-        self.rct_elements = dict(rct_properties['element_counter'])
-        self.rct_sum_moles = rct_properties['sum_moles']
-        self.rct_sum_mass = rct_properties['sum_mass']
-        self.rct_mol_fracs = rct_properties['mol_fracs']
-        self.rct_mass_fracs = rct_properties['mass_fracs']
+        df, elements, sum_moles, sum_masses = self._eq_properties(self._rct_items)
+        self.rct_properties = df
+        self.rct_elements = elements
+        self.rct_moles = sum_moles
+        self.rct_mass = sum_masses
 
     def _assign_prod_attrs(self):
         """
         Assign results for products to attributes.
         """
-        prod_properties = self._equation_properties(self.products)
-        self.prod_species = prod_properties['species']
-        self.prod_moles = prod_properties['moles']
-        self.prod_mw = prod_properties['molwt']
-        self.prod_mass = prod_properties['mass']
-        self.prod_elements = dict(prod_properties['element_counter'])
-        self.prod_sum_moles = prod_properties['sum_moles']
-        self.prod_sum_mass = prod_properties['sum_mass']
-        self.prod_mol_fracs = prod_properties['mol_fracs']
-        self.prod_mass_fracs = prod_properties['mass_fracs']
+        df, elements, sum_moles, sum_masses = self._eq_properties(self._prod_items)
+        self.prod_properties = df
+        self.prod_elements = elements
+        self.prod_moles = sum_moles
+        self.prod_mass = sum_masses
 
     def _check_balance(self):
         """
@@ -236,75 +195,10 @@ class ChemicalEquation:
         tol = 1e-4
         self.balance = True
 
+        if self.rct_elements == self.prod_elements:
+            return
+
         for x, y in zip(self.rct_elements.values(), self.prod_elements.values()):
             if abs(x - y) > tol:
                 self.balance = False
                 break
-
-    def reactant_props(self, species):
-        """
-        Properties for a given reactant.
-
-        Parameters
-        ----------
-        species : str
-            Name of species in reactants.
-
-        Returns
-        -------
-        rct_props : dict
-            Reactant properties defined as `moles`, `mw`, `mass`,
-            `mol_frac`, and `mass_frac` keys.
-
-        Example
-        -------
-        >>> ce = ChemicalEquation('2 HCl + 2 Na -> 2 NaCl + H2')
-        >>> ce.reactant_props('Na')
-        {'moles': 2.0,
-         'mw': 22.99,
-         'mass': 45.98,
-         'mol_frac': 0.5,
-         'mass_frac': 0.3867}
-        """
-        idx = self.rct_species.index(species)
-        moles = self.rct_moles[idx]
-        mw = self.rct_mw[idx]
-        mass = self.rct_mass[idx]
-        mol_frac = self.rct_mol_fracs[idx]
-        mass_frac = self.rct_mass_fracs[idx]
-        rct_props = dict(moles=moles, mw=mw, mass=mass, mol_frac=mol_frac, mass_frac=mass_frac)
-        return rct_props
-
-    def product_props(self, species):
-        """
-        Properties for a given product.
-
-        Parameters
-        ----------
-        species : str
-            Name of species in products.
-
-        Returns
-        -------
-        prod_props : dict
-            Product properties defined as `moles`, `mw`, `mass`,
-            `mol_frac`, and `mass_frac` keys.
-
-        Example
-        -------
-        >>> ce = ChemicalEquation('2 HCl + 2 Na -> 2 NaCl + H2')
-        >>> ce.product_props('NaCl')
-        {'moles': 2.0,
-         'mw': 58.44,
-         'mass': 116.88,
-         'mol_frac': 0.66,
-         'mass_frac': 0.983}
-        """
-        idx = self.prod_species.index(species)
-        moles = self.prod_moles[idx]
-        mw = self.prod_mw[idx]
-        mass = self.prod_mass[idx]
-        mol_frac = self.prod_mol_fracs[idx]
-        mass_frac = self.prod_mass_fracs[idx]
-        prod_props = dict(moles=moles, mw=mw, mass=mass, mol_frac=mol_frac, mass_frac=mass_frac)
-        return prod_props